FIXED #217: Prevent Compount Layer Magnification

+++ indigo-extras/src/indigoextras/ui/components/ScrollPane.scala
@@ -333,7 +333,7 @@ object ScrollPane:
             stack.toBatch.map {
               _.withBlendMaterial(
                 LayerMask(
-                  ctx.parent.bounds.toScreenSpace(ctx.snapGrid * ctx.magnification)
+                  ctx.parent.bounds.toScreenSpace(ctx.snapGrid * ctx.magnification.toInt)
                 )
               )
             }
